Updated MSVC project and fixed MSVC warnings
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@3947 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ src/common/ProjAction.cpp
@@ -69,7 +69,7 @@ CVec Proj_SpawnParent::velocity() const {
 float Proj_SpawnParent::angle() const {
 	switch(type) {
 		case PSPT_NOTHING: return 0;
-		case PSPT_SHOT: return shot->nAngle;
+		case PSPT_SHOT: return (float)shot->nAngle;
 		case PSPT_PROJ: {
 			CVec v = velocity();
 			NormalizeVector(&v);
